+++ db/wpdb.ts
@@ -20,7 +20,7 @@ import { registerExitHandler } from "./cleanup"
 import { RelatedChart } from "site/client/blocks/RelatedCharts/RelatedCharts"
 import { JsonError } from "utils/server/serverUtil"
 import { covidCountryProfileSlug } from "site/server/covid/CovidConstants"
-import { CountryProfileSpec } from "site/client/covid/CovidSearchCountry"
+import { CountryProfileSpec } from "site/client/CountryProfileConstants"
 
 class WPDB {
     conn?: DatabaseConnection
