+++ site/server/sitemap.ts
@@ -5,8 +5,7 @@ import moment from "moment"
 import * as db from "db/db"
 import { countries } from "utils/countries"
 import urljoin from "url-join"
-import { covidCountryProfileRootPath } from "./covid/CovidConstants"
-import { countryProfileTypes } from "site/client/covid/CovidSearchCountry"
+import { countryProfileSpecs } from "site/client/CountryProfileConstants"
 
 interface SitemapUrl {
     loc: string
@@ -45,9 +44,9 @@ export async function makeSitemap() {
 
     urls = urls
         .concat(
-            ...countryProfileTypes.map(profile => {
+            ...countryProfileSpecs.map(spec => {
                 return countries.map(c => ({
-                    loc: urljoin(BAKED_BASE_URL, profile.rootPath, c.slug)
+                    loc: urljoin(BAKED_BASE_URL, spec.rootPath, c.slug)
                 }))
             })
         )
